Updating to check for vsftp. Eventually will check for more

+++ ftp/setup.sh
@@ -1,27 +1,31 @@
 #!/usr/bin/env bash
-#Getting ftpd up and running
-
+#Getting ftp server up and running
+#This works for vsftp currently
 #Asking to see if they have a ftp user
-read -p "Do you have an FTP User? (y/n)" ANS
+read -p "Do you have an FTP User? (y/n): " ANS
 if [ "ANS" == "y" ]; then
-   read -p "What's the user's name?" FTPUSER
+   read -p "User's name: " FTPUSER
 else
    echo "Please make a user FIRST!"
    exit
 fi
 
-#Setting up VSFTPD
-apt-get -y install vsftpd
-rm /etc/vsftpd.conf
-cp ftpconfigfile /etc/vsftpd.conf
-cd /var/run/vsftpd
-#chattr +i vsftp.conf
+#checking to see what version of FTP is installed
+if [ -f "/etc/vsftpd.conf" ]; then
+   echo "Looks like VSFTPD is already installed!"
+   service vsftpd stop
+   
+   #putting the new config file in
+   rm /etc/vsftpd.conf
+   cp ftpconfigfile /etc/vsftpd.conf
+   
+   #managing the FTP dir
+   cd /home
+   rm -r $FTPUSER
+   mkdir $FTPUSER
 
-#making the chroot
-cd /home
-rm -r $FTPUSER
-mkdir $FTPUSER
-chmod a-w /$FTPUSER
+   #chmod a-w /$FTPUSER
+fi
 
 #restarting
 service vsftpd restart
